+++ application/shared/controller/timemanager/jobs.php
@@ -2,7 +2,7 @@
 /**
  * Author: Kenyon Haliwell
  * Date Created: 12/09/13
- * Date Modified: 1/8/14
+ * Date Modified: 1/9/14
  * Purpose: Controller for Jobs/Job Tracking
  */
 
@@ -93,6 +93,26 @@ class timemanager_jobs extends controller {
     }
     
     /**
+     * Purpose: Used to view/modify the quote of a job
+     */
+    public function quote($job_uid) {
+        $this->load_dependencies(array('loggedIn', 'renderPage', 'jobs', 'departments', 'quotes', 'settings'));
+        $this->sys->template->job_uid = $job_uid;
+        
+        if ($this->is_logged_in()) {
+            $this->sys->template->admin                 = $this->model_settings->is_admin();
+            $this->sys->template->quote                 = $this->model_quotes->get($job_uid);
+            $this->sys->template->departments           = $this->model_departments->get(true);
+            $this->sys->template->title                 = 'Time Manager | Jobs | Quote';
+            $this->sys->template->jobs_quoting_active   = 'class="active"';
+            
+            $this->model_renderPage->parse('jobs_quote', true);
+        } else {
+            $this->load_home();
+        }
+    }
+    
+    /**
      * Purpose: Loads page that has information about job tracking
      */
     public function tracking($page_id=1) {
@@ -125,7 +145,7 @@ class timemanager_jobs extends controller {
      * Purpose: Load add jobs page
      */
     public function add() {
-        $this->load_dependencies(array('loggedIn', 'renderPage', 'settings', 'clients', 'jobs', 'departments'));
+        $this->load_dependencies(array('loggedIn', 'renderPage', 'settings', 'clients', 'jobs', 'departments', 'quotes'));
 
         if (!isset($this->sys->template->response)) {
             $this->sys->template->response = '';
@@ -142,6 +162,10 @@ class timemanager_jobs extends controller {
             $this->sys->template->title                 = 'Time Manager | Jobs | Add';
             $this->sys->template->jobs_quoting_active   = 'class="active"';
             $this->sys->template->clients               = $this->model_clients->get();
+            
+            if (array_key_exists('job_id', $_POST)) {
+                $this->model_quotes->add();
+            }
             
             $parse = 'jobs_add';
             $full_page = True;
@@ -213,25 +237,6 @@ class timemanager_jobs extends controller {
 
         //Parses the HTML from the view
         $this->model_renderPage->parse($parse, $full_page);
-    }
-    
-    /**
-     * Purpose: Used to view/modify the quote of a job
-     */
-    public function quote($job_uid) {
-        $this->load_dependencies(array('loggedIn', 'renderPage', 'jobs', 'settings'));
-        $this->sys->template->job_uid = $job_uid;
-        
-        if ($this->is_logged_in()) {
-            $this->sys->template->admin                 = $this->model_settings->is_admin();
-            $this->sys->template->job                   = $this->model_jobs->get($job_uid, false);
-            $this->sys->template->title                 = 'Time Manager | Jobs | Quote';
-            $this->sys->template->jobs_quoting_active   = 'class="active"';
-            
-            $this->model_renderPage->parse('jobs_quote', true);
-        } else {
-            $this->load_home();
-        }
     }
     
     /**