+++ timemanagerAssets/js/main.js
@@ -1,6 +1,6 @@
 var web_root = '/timemanager/timemanager/';
-
-jQuery(document).ready(function() {   
+    
+jQuery(document).ready(function() {
     jQuery('input[type=submit]').button();
     
     jQuery('.date').datepicker({
@@ -12,4 +12,9 @@ jQuery(document).ready(function() {
 
     jQuery('.date, .inline_date').datepicker('option', 'dateFormat', 'mm/dd/y');
     jQuery('.date, .inline_date').datepicker('option', 'altFormat', 'mm/dd/y');
+    
+    if (jQuery('#response').length == 1) {
+        var response = document.getElementById('response');
+        response.scrollIntoViewIfNeeded();
+    }
 });